backend/requirements.txt
fastapi
uvicorn
motor
pymongo
pydantic
pydantic-settings
python-dotenv
python-multipart
scikit-learn
pandas
numpy
google-auth
